+++ database/migrations/2019_09_29_113444_create_expenses_table.php
@@ -18,7 +18,7 @@ class CreateExpensesTable extends Migration
             $table->string('description');
             $table->integer('price');
             $table->string('note')->nullable();
-            $table->unsignedBigInteger('user_id')->nullable()->unsigned();
+            $table->BigInteger('user_id')->nullable()->unsigned();
             $table->softDeletes();
             $table->timestamps();
         });
